Fixed reconnection bug
This fix adds an optional network (WiFi) and client (MQTTT) connection algorithm in EspMQTTClient::loop(). By add, I mean that it puts the new code/algorithm behind a #ifdef WIFI_FIX blocks so that the person testing the code can switch between the new and old algorithms.

The new algorithm uses a state machine formulation as opposed to the prior condition-styled algorithm. The state machine moves through six states that are documented in the comments in EspMQTTClient::loop().

Fix developed by Michael McDonald of Albuquerque NM.